Move aux-images component
