+++ apps/chat/components/aux-images.tsx
